feat(web-app): place Tunings page under /maintenance
