+++ packages/cube-frontend-web-app/src/pages/events/links.ts
@@ -3,6 +3,5 @@ import { CosRoutesEnum } from '@cube-frontend/web-app/enum/routes'
 export const links = {
   events: CosRoutesEnum.EVENTS_PAGE,
   triggers: CosRoutesEnum.EVENTS_TRIGGERS_PAGE,
-  tunings: CosRoutesEnum.EVENTS_TUNINGS_PAGE,
   chart: CosRoutesEnum.EVENTS_CHART_PAGE,
 }
