+++ packages/cube-frontend-web-app/src/pages/maintenance/links.ts
@@ -2,5 +2,6 @@ import { CosRoutesEnum } from '@cube-frontend/web-app/enum/routes'
 
 export const links = {
   supportFiles: CosRoutesEnum.MAINTENANCE_SUPPORT_FILES_PAGE,
+  tunings: CosRoutesEnum.MAINTENANCE_TUNINGS_PAGE,
   license: CosRoutesEnum.MAINTENANCE_LICENSE_PAGE,
 }
